# Conflicts:
#	image.Makefile
#	pkg/microservice/aslan/core/common/repository/mongodb/service.go
#	pkg/microservice/aslan/core/common/service/service.go
#	pkg/microservice/aslan/core/workflow/service/webhook/syncer.go

+++ pkg/microservice/aslan/core/common/repository/models/s3.go
@@ -32,7 +32,7 @@ type S3Storage struct {
 	EncryptedSk string             `bson:"encryptedSk"    json:"-"`
 	UpdatedBy   string             `bson:"updated_by"     json:"updated_by"`
 	UpdateTime  int64              `bson:"update_time"    json:"update_time"`
-	Provider    int32              `bson:"provider"       json:"provider"`
+	Provider    int8               `bson:"provider"       json:"provider"`
 }
 
 func (s S3Storage) TableName() string {
